Got code to check that the requested branch exists.
The branching code now does a separate clone then a checkout in the
cloned repository.  The original attempt at cloning the branch with
'git clone --branch' will still checkout the current head if the
entered branch didn't exists and the command still returned a 0
(success return code).  With an initial 'git clone' and a separate
'git checkout' allows us to detect if the checkout failed; the checkout
now returns a '1' if you enter a branch that doesn't exist, allowing
the export to fail gracefully without performing the export.